refactor: extract function to modularize
